requirements.txt
streamlit
streamlit-calendar
pandas
google-genai
python-dotenv